Merge pull request #17 from DylanCheetah/master (with fixes)
Fixed a bug that causes an application to crash if the underscore char
is typed. (sic)
The bug was caused when the event string is "Shift--" because splitting
that string results in a list of 3 strings instead of the expected list
of 2 strings.

Use str.partition instead of str.split with maxsplit=1.